% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.table.R
\name{conf.table}
\alias{conf.table}
\title{Configuration Table}
\usage{
conf.table(data, ncut = ncut)
}
\arguments{
\item{data}{name of the model object; the table of solutions for an application of QCA. Default set to \code{data}.}

\item{ncut}{configurational n levels for inclusion. Default set to \code{ncut=4}}
}
\value{
The output of the Bootstrapped Recommendation
#' @export
}
\description{
Internal function; calculates via logistic regression the output of the Bootstrapped Robustness Recommendation
}
